RAG/lab2/requirements.txt
langchain
langgraph
openai
streamlit
tavily-python
python-dotenv
